requirements.txt
fastapi
uvicorn
requests
jinja2
python-docx
reportlab
pydantic
python-dotenv
lxml

